Remove unused classes (#804)
* Remove unused classes

* Remove dead code

* Remove NavigatingActionWaitOptions

* Remove more classes

* More formats

* Ops

* MarginOptions to Margin

* Remove PageCloseOptions

* Remove weird file

* Fix tests

* Change double to decimal

+++ src/PlaywrightSharp/DeviceDescriptor.cs
@@ -33,7 +33,7 @@ namespace PlaywrightSharp
         /// <summary>
         /// Device scale factor.
         /// </summary>
-        public double DeviceScaleFactor { get; set; }
+        public decimal DeviceScaleFactor { get; set; }
 
         /// <summary>
         /// Converts the <see cref="DeviceDescriptor"/> to <see cref="BrowserContextOptions"/>.
